refactor(#425): extract usePanelResize hook, deduplicate panel resize logic (#459)
* refactor(#425): extract usePanelResize hook, deduplicate panel resize logic

- Create hooks/usePanelResize.ts: generic panel resize hook
  (mousedown/move/up, cursor style, min/max clamping, cleanup)
- Migrate Sidebar.tsx: replace 49 lines of inline resize code
- Migrate ChatConsole.tsx: replace 41 lines of inline resize code
- Both use same hook with different computeWidth strategies:
  Sidebar: e.clientX - rect.left (left-mounted)
  ChatConsole: window.innerWidth - e.clientX (right-mounted)

Closes #425

* chore: trigger CI re-run

* chore: trigger CI re-run (PR template already fixed)

---------
